backend-starter/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-filter>=24.2
python-decouple>=3.8
gunicorn>=21.2
psycopg2-binary>=2.9
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
dj-database-url>=2.2
Pillow>=10.0
celery>=5.3
requests>=2.31
croniter>=2.0
whitenoise>=6.6
python-json-logger>=2.0
locust>=2.29
